pyproject.toml
[project]
name = "swsec-intro-backend-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi-cli>=0.0.10",
    "fastapi[standard]>=0.116.0",
    "pyjwt==2.3.0",
]
